feat: move fields to global widgets folder
