+++ admin/lib/posts/view/post_write.dart
@@ -7,8 +7,7 @@
 import 'package:admin/posts/models/post.dart';
 import 'package:admin/posts/state/post_bloc.dart';
 import 'package:admin/posts/view/post_preview.dart';
-import 'package:admin/posts/view/widgets/editable_image_field.dart';
-import 'package:admin/posts/view/widgets/editable_tile.dart';
+import 'package:admin/widgets/fields.dart';
 import 'package:admin/posts/view/widgets/html_tag_editor.dart';
 import 'package:admin/widgets/loadings.dart';
 import 'package:admin/widgets/utils.dart';
